Automatic Documentations

Photon.ADADelta — Type

ADADelta(ρ = 0.9)

ADADelta is a version of ADAGrad adapting its learning rate based on a window of past gradient updates. Parameters don't need tuning.

Parameters

Rho (ρ): Factor by which the gradient is decayed at each time step.

Examples

opt = ADADelta()
opt = ADADelta(0.89)

Photon.ADAGrad — Type

ADAGrad(η = 0.1)

ADAGrad optimizer. It has parameter specific learning rates based on how frequently it is updated. Parameters don't need tuning.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Examples

opt = ADAGrad()
opt = ADAGrad(0.001)

Photon.ADAM — Type

ADAM(η = 0.001, β::Tuple = (0.9, 0.999))

ADAM optimiser.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Decay of momentums (β::Tuple): Exponential decay for the first (β1) and the second (β2) momentum estimate.

Examples

opt = ADAM()
opt = ADAM(0.001, (0.9, 0.8))

Photon.AMSGrad — Type

AMSGrad(η = 0.001, β::Tuple = (0.9, 0.999))

The AMSGrad version of the ADAM optimiser. Parameters don't need tuning.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Decay of momentums (β::Tuple): Exponential decay for the first (β1) and the second (β2) momentum estimate.

Examples

opt = AMSGrad()
opt = AMSGrad(0.001, (0.89, 0.995))

Photon.AdaBelief — Type

AdaBelief(η = 0.001, β::Tuple = (0.9, 0.999))

The AdaBelief optimiser is a variant of the well-known ADAM optimiser.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Decay of momentums (β::Tuple): Exponential decay for the first (β1) and the second (β2) momentum estimate.

Examples

opt = AdaBelief()
opt = AdaBelief(0.001, (0.9, 0.8))

Photon.AdaMax — Type

AdaMax(η = 0.001, β::Tuple = (0.9, 0.999))

AdaMax is a variant of ADAM based on the ∞-norm.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Decay of momentums (β::Tuple): Exponential decay for the first (β1) and the second (β2) momentum estimate.

Examples

opt = AdaMax()
opt = AdaMax(0.001, (0.9, 0.995))

Photon.Descent — Type

Descent(η = 0.1)

Classic gradient descent optimiser with learning rate η. For each parameter p and its gradient δp, this runs p -= η*δp

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Photon.Momentum — Type

Momentum(η = 0.01, ρ = 0.9)

Gradient descent optimizer with learning rate η and momentum ρ.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Momentum (ρ): Controls the acceleration of gradient descent in the prominent direction, in effect dampening oscillations.

Examples

opt = Momentum()
opt = Momentum(0.01, 0.99)

Photon.NADAM — Type

NADAM(η = 0.001, β::Tuple = (0.9, 0.999))

NADAM is a Nesterov variant of ADAM. Parameters don't need tuning.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Decay of momentums (β::Tuple): Exponential decay for the first (β1) and the second (β2) momentum estimate.

Examples

opt = NADAM()
opt = NADAM(0.002, (0.89, 0.995))

Photon.Nesterov — Type

Nesterov(η = 0.001, ρ = 0.9)

Gradient descent optimizer with learning rate η and Nesterov momentum ρ.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Nesterov momentum (ρ): Controls the acceleration of gradient descent in the prominent direction, in effect dampening oscillations.

Examples

opt = Nesterov()
opt = Nesterov(0.003, 0.95)

Photon.OADAM — Type

OADAM(η = 0.0001, β::Tuple = (0.5, 0.9))

OADAM (Optimistic ADAM) is a variant of ADAM adding an "optimistic" term suitable for adversarial training.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Decay of momentums (β::Tuple): Exponential decay for the first (β1) and the second (β2) momentum estimate.

Examples

opt = OADAM()
opt = OADAM(0.001, (0.9, 0.995))

Photon.RADAM — Type

RADAM(η = 0.001, β::Tuple = (0.9, 0.999))

Rectified ADAM optimizer.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Decay of momentums (β::Tuple): Exponential decay for the first (β1) and the second (β2) momentum estimate.

Examples

opt = RADAM()
opt = RADAM(0.001, (0.9, 0.8))

Photon.RMSProp — Type

RMSProp(η = 0.001, ρ = 0.9)

Optimizer using the RMSProp algorithm. Often a good choice for recurrent networks. Parameters other than learning rate generally don't need tuning.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Momentum (ρ): Controls the acceleration of gradient descent in the prominent direction, in effect dampening oscillations.

Examples

opt = RMSProp()
opt = RMSProp(0.002, 0.95)

Photon.ADAMW — Function

ADAMW(η = 0.001, β::Tuple = (0.9, 0.999), decay = 0)

ADAMW is a variant of ADAM fixing (as in repairing) its weight decay regularization.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Decay of momentums (β::Tuple): Exponential decay for the first (β1) and the second (β2) momentum estimate.

decay: Decay applied to weights during optimisation.

Examples

opt = ADAMW()
opt = ADAMW(0.001, (0.89, 0.995), 0.1)

Photon.Coordinate — Method

Coordinate(g::Grid ,::GridType{:Dual},Ind::CartesianIndex)

Returns the equivalent dual coordinate of of a CartesianIndex.

The function doesnot chack for being inside the grid.

Photon.Coordinate — Method

Coordinate(g::Grid , ::GridType{:Primal}, Ind::CartesianIndex)

Returns the equivalent primal coordinate of of a CartesianIndex.

The function doesnot chack for being inside the grid.

Photon.Coordinates — Method

 Coordinates(g::Grid,gridtype::GridType{T},CIns::CartesianIndices ) where T

Returns the primal or dual coordinates of a set of cartesian indecies.

Photon.Coordinates — Method

  Coordinates(g::Grid,gridtype::GridType{T}) where T

Returns the primal or dual coordinates of all points on a grid.