Automatic Documentations

Base.IteratorsMD.CartesianIndex — Method

CartesianIndex(g::Grid{ND},::GridType{:Dual},  coordinate::NTuple{ND,Number}) where ND

Returns the equivalent dualCartesianIndex of a coordinate.

Photon.PML — Method

 PML(g::Grid{ND})

Returns an object to be used later as functor that can be used to customize the PML length, sides and directions.

Photon.SFRegion — Method

SFRegion(g::Grid{ND},region::Function = (x-> false))

Returns a ScatteredField region defined by the boolean function region.

Photon.Coordinate — Method

Coordinate(g::Grid ,::GridType{:Dual},Ind::CartesianIndex)

Returns the equivalent dual coordinate of of a CartesianIndex.

The function doesnot chack for being inside the grid.

Photon.Coordinate — Method

Coordinate(g::Grid , ::GridType{:Primal}, Ind::CartesianIndex)

Returns the equivalent primal coordinate of of a CartesianIndex.

The function doesnot chack for being inside the grid.

Photon.Coordinates — Method

 Coordinates(g::Grid,gridtype::GridType{T},CIns::CartesianIndices ) where T

Returns the primal or dual coordinates of a set of cartesian indecies.

Photon.Coordinates — Method

  Coordinates(g::Grid,gridtype::GridType{T}) where T

Returns the primal or dual coordinates of all points on a grid.

Photon.eigensolve — Method

eigensolve(dev::Device{2}; direction::Direction,center::Tuple , k::Real)

solves the eigenvalue problem across the line noraml to direction and centered at center.

The method was created to be used for defining mode sources for 2D problems.